Important features
This project is pretty easy and simple to follow, you can learn how to use Jetpack Compose with some other libraries in order to manage some basics operations with the users. What you can do with this project? You can use this example code to recreate some kind of splash screen logic or other preferences for the user.
The project uses:
- MVVM Architecture
- Jectpack Compose
- Data Store Preferences
- Light and dark theming
- Coroutines
- Compose States